requirements.txt
aiofiles
aiohttp
asyncio
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
lyricsgenius
motor
pillow
psutil
py-tgcalls==0.9.7
pykeyboard
pyrogram==2.0.106
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
yt-dlp
uvloop